requirements.txt
Django==4.0.3
djangorestframework==3.13.1
djoser==2.1.0
psycopg2-binary==2.9.3
python-decouple==3.6
gunicorn==20.1.0
django-heroku==0.3.1
dj-database-url==0.5.0